+++ Library/Aphid/Support/Compatibility.js
@@ -7,5 +7,6 @@
 
 Aphid.Support.Compatibility = {};
 
+//= require <Aphid/Support/Compatibility/Window.js>
 //= require <Aphid/Support/Compatibility/String>
 //= require <Aphid/Support/Compatibility/HTML5>