+++ Library/Aphid/Support/Compatibility/Window.js
@@ -1,0 +1,14 @@
+/**
+ * Aphid.Support.Compatibility.Window
+**/
+
+if (Object.isUndefined(window.stop))
+{
+  window.stop = function()
+  {
+    try {
+      document.executeCommand("Stop");
+    }
+    catch(err) {}
+  }
+}